+++ normalized-cache-incubating/src/commonMain/kotlin/com/apollographql/cache/normalized/api/CacheResolver.kt
@@ -5,6 +5,7 @@ import com.apollographql.apollo.api.Executable
 import com.apollographql.apollo.api.MutableExecutionOptions
 import com.apollographql.apollo.exception.CacheMissException
 import com.apollographql.apollo.mpp.currentTimeMillis
+import com.apollographql.cache.normalized.api.CacheResolver.ResolvedValue
 import com.apollographql.cache.normalized.maxStale
 import com.apollographql.cache.normalized.storeExpirationDate
 import com.apollographql.cache.normalized.storeReceiveDate
@@ -68,9 +69,14 @@ interface CacheResolver {
    * @param context the field to resolve and associated information to resolve it
    *
    * @return a value that can go in a [Record]. No type checking is done. It is the responsibility of implementations to return the correct
-   * type
+   * type. The value can be wrapped in a [ResolvedValue] to provide additional information.
    */
   fun resolveField(context: ResolverContext): Any?
+
+  class ResolvedValue(
+      val value: Any?,
+      val cacheHeaders: CacheHeaders,
+  )
 }
 
 class ResolverContext(
@@ -133,7 +139,7 @@ object DefaultCacheResolver : CacheResolver {
 
 /**
  * A cache resolver that raises a cache miss if the field's received date is older than its max age
- * (configurable via [maxAgeProvider]) or its expiration date has passed.
+ * (configurable via [maxAgeProvider]) or if its expiration date has passed.
  *
  * Received dates are stored by calling `storeReceiveDate(true)` on your `ApolloClient`.
  *
@@ -145,51 +151,62 @@ object DefaultCacheResolver : CacheResolver {
  * @see MutableExecutionOptions.storeExpirationDate
  * @see MutableExecutionOptions.maxStale
  */
-class ExpirationCacheResolver(
+class CacheControlCacheResolver(
     private val maxAgeProvider: MaxAgeProvider,
 ) : CacheResolver {
   /**
-   * Creates a new [ExpirationCacheResolver] with no max ages. Use this constructor if you want to consider only the expiration dates.
+   * Creates a new [CacheControlCacheResolver] with no max ages. Use this constructor if you want to consider only the expiration dates.
    */
   constructor() : this(maxAgeProvider = GlobalMaxAgeProvider(Duration.INFINITE))
 
   override fun resolveField(context: ResolverContext): Any? {
-    val resolvedField = FieldPolicyCacheResolver.resolveField(context)
+    var isStale = false
     if (context.parent is Record) {
       val field = context.field
-      val maxStale = context.cacheHeaders.headerValue(ApolloCacheHeaders.MAX_STALE)?.toLongOrNull() ?: 0L
-      val currentDate = currentTimeMillis() / 1000
-
-      // Consider the field's max age (client side)
-      val fieldMaxAge = maxAgeProvider.getMaxAge(MaxAgeContext(context.path)).inWholeSeconds
-      val fieldReceivedDate = context.parent.receivedDate(field.name)
-      if (fieldReceivedDate != null) {
-        val fieldAge = currentDate - fieldReceivedDate
-        val stale = fieldAge - fieldMaxAge
-        if (stale >= maxStale) {
+      val receivedDate = context.parent.receivedDate(field.name)
+      // Consider the client controlled max age
+      if (receivedDate != null) {
+        val currentDate = currentTimeMillis() / 1000
+        val age = currentDate - receivedDate
+        val maxAge = maxAgeProvider.getMaxAge(MaxAgeContext(context.path)).inWholeSeconds
+        val staleDuration = age - maxAge
+        val maxStale = context.cacheHeaders.headerValue(ApolloCacheHeaders.MAX_STALE)?.toLongOrNull() ?: 0L
+        if (staleDuration >= maxStale) {
           throw CacheMissException(
-              context.parentKey,
-              context.fieldKeyGenerator.getFieldKey(FieldKeyContext(context.parentType, context.field, context.variables)),
-              true
+              key = context.parentKey,
+              fieldName = context.fieldKeyGenerator.getFieldKey(FieldKeyContext(context.parentType, context.field, context.variables)),
+              stale = true
           )
         }
+        if (staleDuration >= 0) isStale = true
       }
 
-      // Consider the field's expiration date (server side)
-      val fieldExpirationDate = context.parent.expirationDate(field.name)
-      if (fieldExpirationDate != null) {
-        val stale = currentDate - fieldExpirationDate
-        if (stale >= maxStale) {
+      // Consider the server controlled max age
+      val expirationDate = context.parent.expirationDate(field.name)
+      if (expirationDate != null) {
+        val currentDate = currentTimeMillis() / 1000
+        val staleDuration = currentDate - expirationDate
+        val maxStale = context.cacheHeaders.headerValue(ApolloCacheHeaders.MAX_STALE)?.toLongOrNull() ?: 0L
+        if (staleDuration >= maxStale) {
           throw CacheMissException(
-              context.parentKey,
-              context.fieldKeyGenerator.getFieldKey(FieldKeyContext(context.parentType, context.field, context.variables)),
-              true
+              key = context.parentKey,
+              fieldName = context.fieldKeyGenerator.getFieldKey(FieldKeyContext(context.parentType, context.field, context.variables)),
+              stale = true
           )
         }
+        if (staleDuration >= 0) isStale = true
       }
     }
 
-    return resolvedField
+    val value = FieldPolicyCacheResolver.resolveField(context)
+    return if (isStale) {
+      ResolvedValue(
+          value = value,
+          cacheHeaders = CacheHeaders.Builder().addHeader(ApolloCacheHeaders.STALE, "true").build(),
+      )
+    } else {
+      value
+    }
   }
 }