+++ tests/pagination/src/commonTest/kotlin/ConnectionWithNodesPaginationTest.kt
@@ -63,7 +63,7 @@ class ConnectionWithNodesPaginationTest {
       }
     }
     apolloStore.writeOperation(query1, data1)
-    var dataFromStore = apolloStore.readOperation(query1)
+    var dataFromStore = apolloStore.readOperation(query1).data
     assertEquals(data1, dataFromStore)
     assertChainedCachesAreEqual(apolloStore)
 
@@ -86,7 +86,7 @@ class ConnectionWithNodesPaginationTest {
       }
     }
     apolloStore.writeOperation(query2, data2)
-    dataFromStore = apolloStore.readOperation(query1)
+    dataFromStore = apolloStore.readOperation(query1).data
     var expectedData = UsersQuery.Data {
       users = buildUserConnection {
         pageInfo = buildPageInfo {
@@ -131,7 +131,7 @@ class ConnectionWithNodesPaginationTest {
       }
     }
     apolloStore.writeOperation(query3, data3)
-    dataFromStore = apolloStore.readOperation(query1)
+    dataFromStore = apolloStore.readOperation(query1).data
     expectedData = UsersQuery.Data {
       users = buildUserConnection {
         pageInfo = buildPageInfo {
@@ -182,7 +182,7 @@ class ConnectionWithNodesPaginationTest {
       }
     }
     apolloStore.writeOperation(query4, data4)
-    dataFromStore = apolloStore.readOperation(query1)
+    dataFromStore = apolloStore.readOperation(query1).data
     expectedData = UsersQuery.Data {
       users = buildUserConnection {
         pageInfo = buildPageInfo {
@@ -239,7 +239,7 @@ class ConnectionWithNodesPaginationTest {
       }
     }
     apolloStore.writeOperation(query5, data5)
-    dataFromStore = apolloStore.readOperation(query1)
+    dataFromStore = apolloStore.readOperation(query1).data
     assertEquals(data5, dataFromStore)
     assertChainedCachesAreEqual(apolloStore)
 
@@ -256,7 +256,7 @@ class ConnectionWithNodesPaginationTest {
       }
     }
     apolloStore.writeOperation(query6, data6)
-    dataFromStore = apolloStore.readOperation(query1)
+    dataFromStore = apolloStore.readOperation(query1).data
     assertEquals(data5, dataFromStore)
     assertChainedCachesAreEqual(apolloStore)
   }